+++ src/api/index.ts
@@ -1,0 +1,7 @@
+import Api from './Api';
+import MobileApi from './MobileApi';
+
+export {
+    Api,
+    MobileApi,
+};