+++ src/constants/privacy.js
@@ -1,5 +1,0 @@
-// @flow strict
-
-export const PRIVACY_EVERYONE: 0 = 0;
-export const PRIVACY_FRIENDS: 1 = 1;
-export const PRIVACY_ME: 2 = 2;