+++ src/exceptions/EndomondoException.js
@@ -1,6 +1,0 @@
-// @flow strict
-export default class EndomondoException extends Error {
-    constructor(message: string) {
-        super(`Endomondo Error: ${message}`);
-    }
-}