+++ src/factories/index.js
@@ -1,8 +1,0 @@
-// @flow strict
-import PointFactory from './PointFactory';
-import WorkoutFactory from './WorkoutFactory';
-
-export {
-    PointFactory,
-    WorkoutFactory,
-};