+++ src/models/index.js
@@ -1,8 +1,0 @@
-// @flow strict
-import Point from './Point';
-import Workout from './Workout';
-
-export {
-    Point,
-    Workout,
-};