+++ src/types/api/Paging.ts
@@ -1,0 +1,7 @@
+export type Paging = {
+    paging: {
+        total: number,
+        next: string,
+        previous?: string,
+    },
+};