+++ src/types/api/Point.ts
@@ -1,0 +1,14 @@
+export type Point = {
+    time: string,
+    instruction?: number,
+    latitude: number,
+    longitude: number,
+    distance: number,
+    duration: number,
+    altitude: number,
+    sensor_data: {
+        speed?: number,
+        heart_rate?: number,
+        cadence?: number,
+    },
+};