+++ src/types/api/Workouts.js
@@ -1,7 +1,0 @@
-// @flow strict
-import type { Workout } from './Workout';
-import type { Paging } from './Paging';
-
-export type Workouts = {
-    data: Array<Workout>,
-} & Paging;